Add hazard unit, refactor support.py, fix issue with branch
